+++ plugins/tengine-web-server.rb
@@ -8,8 +8,9 @@ Plugin.define do
 name "Tengine-Web-Server"
 authors [
   "Brendan Coles <bcoles@gmail.com>", # 2012-05-17
+  "Andrew Horton", # v0.2 # 2025-08-02 # Added Alibaba-specific Tengine header detection
 ]
-version "0.1"
+version "0.2"
 description "Tengine is a web server originated by Taobao, the largest e-commerce website in Asia. It is based on the popular Nginx HTTP server."
 website "http://tengine.taobao.org/"
 
@@ -26,6 +27,11 @@ matches [
 
 # HTTP Server Header # Version Detection
 { :search=>"headers[server]", :version=>/^Tengine\/([^\s]+)/ },
+
+# Alibaba-specific headers often found with Tengine
+{ :search=>"headers[eagleid]", :name=>"Alibaba EagleId Header" },
+{ :search=>"headers[eagleeye-traceid]", :name=>"Alibaba EagleEye Trace ID" },
+{ :search=>"headers[timing-allow-origin]", :name=>"Tengine Timing Allow Origin" },
 
 ]